[REF] 글 추가 기능 서버 연동 완료

+++ HotspoTalk/app/src/main/java/com/example/hotspotalk/data/entity/repuest/CreateRoom.kt
@@ -4,8 +4,8 @@ data class CreateRoom(
     val name: String?,
     val password: String?,
     val memberLimit: Int?,
-    val latitude: Int?,
-    val longitude: Int?,
-    val areaType: Int?,
-    val areaDetail: Int?
+    val latitude: Double?,
+    val longitude: Double?,
+    val areaType: Int?, // 0일때 반경 1일때 주소
+    val areaDetail: Int? // 미터
 )
